Solve remove photo button problems from mobgenners edit

+++ backend/controllers/MobgennersController.php
@@ -293,12 +293,17 @@ class MobgennersController extends Controller
 
             $model = $this->findModel($mobgennerId);
             if ($model) {
-                $filename = Yii::$app->params["BACKEND_WEB"] . "files/mobgenners/" . $model->image;
-                unlink($filename);
-                $model->image = '';
-                $model->save();
-                $data = array('1' => 'Done');
-                return json_encode($data);
+                if ($model->image != '') {
+                    $filename = Yii::$app->params["BACKEND_WEB"] . "files/mobgenners/" . $model->image;
+                    unlink($filename);
+                    $model->image = '';
+                    $model->save();
+                    $data = array('1' => 'Done');
+                    return json_encode($data);
+                } else {                    
+                    $data = array('1' => 'Nothing to do.');
+                    return json_encode($data);
+                }
             }
     }
